+++ neurons/utils/__init__.py
@@ -6,6 +6,7 @@ from .pre_flight import (
 from .system import restart_app, clean_temp_files
 from .auto_update import AutoUpdate
 from . import wandb_logger
+from .rate_limiter import with_rate_limit
 
 __all__ = [
     "run_shared_preflight_checks",
@@ -15,4 +16,5 @@ __all__ = [
     "clean_temp_files",
     "AutoUpdate",
     "wandb_logger",
+    "with_rate_limit",
 ]
